Fixed Energy Beam has no damage if not fully charged

+++ src/main/java/com/mishkapp/minecraft/plugins/squarekit/suffixes/use/EnergyBeam.java
@@ -183,7 +183,7 @@ public class EnergyBeam extends UseSuffix {
         Entity target = getTarget();
 
         double distance = getDistance();
-        double damage = (lastCharge/ maxCharge) * maxDamage;
+        double damage = (lastCharge/ (double)maxCharge) * maxDamage;
         if(target != null){
             Vector3d headshotPoint = isHeadshot(target);
             boolean headshot = headshotPoint != null;
